[Update] Implemented requested changes for protocol & code quality
